requirements.txt
flowlauncher
python-dotenv
requests
typing